pyproject.toml
[project]
name = "pavement-detect"
version = "1.0.0"
description = "同济大学SITP项目：基于视频图像的路面病害识别与展示"
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = { text = "MIT" }
authors = [
    { name = "Tongji SITP Team" },
]

dependencies = [
    "streamlit>=1.30.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "Pillow>=10.0.0",
    "requests>=2.31.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "PyYAML>=6.0",
    "matplotlib>=3.7.0",
    "scipy>=1.10.0",
    "tqdm>=4.65.0",
    "seaborn>=0.12.0",
    "psutil>=5.9.0",
    "py-cpuinfo>=9.0.0",
    "thop>=0.1.1",
    "timm==0.9.16",
    "einops>=0.6.0",
    "efficientnet_pytorch>=0.7.0",
    "ultralytics>=8.1.0",
    "onnxruntime<=1.19.2",
    "pycocotools>=2.0.6",
    "tensorboard>=2.13.0",
    "dill>=0.3.0",
    "reportlab>=4.0.0",
    "python-docx>=0.8.11",
]

[project.scripts]
pavement-web = "pavement_detect.app:main"
pavement-train = "pavement_detect.train:main"
pavement-plot = "pavement_detect.plot_academic:generate_academic_plots"

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }

[tool.uv]
prerelease = "allow"
environments = ["sys_platform == 'win32'"]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/nightly/cu126"
explicit = true


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pavement_detect"]
